+++ spec/pio/open_flow10/hello_failed_spec.rb
@@ -1,0 +1,25 @@
+require 'pio/open_flow10/hello_failed'
+
+describe Pio::OpenFlow10::HelloFailed do
+  it_should_behave_like('an OpenFlow message', Pio::OpenFlow10::HelloFailed)
+
+  describe '.new' do
+    When(:hello_failed) { Pio::OpenFlow10::HelloFailed.new(options) }
+
+    context 'with {}' do
+      Given(:options) { {} }
+
+      Then { hello_failed.length == 12 }
+      Then { hello_failed.error_type == :hello_failed }
+      Then { hello_failed.error_code == :incompatible }
+      Then { hello_failed.description == '' }
+    end
+
+    context "with description: 'error description'" do
+      Given(:options) { { description: 'error description' } }
+
+      Then { hello_failed.length == 29 }
+      Then { hello_failed.description == 'error description' }
+    end
+  end
+end